:zap: 결제 도메인 포인트 도메인과 분리
