+++ src/main/java/com/hhplus/concertReserv/application/PaymentFacade.java
@@ -1,37 +1,39 @@
 package com.hhplus.concertReserv.application;
 
 import com.hhplus.concertReserv.domain.member.dto.PointDto;
-import com.hhplus.concertReserv.domain.member.service.PointService;
+import com.hhplus.concertReserv.domain.reservation.service.PaymentService;
 import com.hhplus.concertReserv.domain.token.service.TokenService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.orm.ObjectOptimisticLockingFailureException;
 import org.springframework.stereotype.Component;
 
-import java.util.UUID;
-
 @Component
 @Slf4j
-public class PointFacade {
+public class PaymentFacade {
 
-    private final PointService pointService;
+    private final PaymentService paymentService;
+
     private final TokenService tokenService;
 
-
-    public PointFacade(PointService pointService, TokenService tokenService) {
-        this.pointService = pointService;
+    public PaymentFacade(PaymentService paymentService, TokenService tokenService) {
+        this.paymentService = paymentService;
         this.tokenService = tokenService;
     }
 
-
-    public PointDto charge(PointCommand.Charge requestBody) {
+    public PointDto paid(PointCommand.Paid requestBody) {
         PointDto result = new PointDto();
-        boolean chargeSuccess = false;
+
+        boolean paidSuccess = false;
         int retryCount = 0;
-        while (!chargeSuccess & retryCount < 10000) {
+
+        while (!paidSuccess && retryCount < 10000) {
             retryCount += 1;
             try {
-                result = pointService.charge(requestBody.memberId(), requestBody.amount());
-                chargeSuccess = true;
+                result = paymentService.paid(requestBody.paymentId(), requestBody.amount());
+                paidSuccess = true;
+                // 토큰 만료화
+                tokenService.deactivateToken(result.getMemberId(), result.getConcertId());
+
             } catch (ObjectOptimisticLockingFailureException e) {
                 log.error(e.toString());
                 log.info(String.format("======= retryCount : %d, amount : %d ======", retryCount, requestBody.amount()));
@@ -39,15 +41,7 @@ public class PointFacade {
                 log.error(e.toString());
                 break;
             }
-
         }
-
         return result;
-    }
-
-
-
-    public PointDto getPoint(UUID memberId) {
-        return pointService.getPoint(memberId);
     }
 }
