+++ src/main/java/com/hhplus/concertReserv/domain/member/service/PointService.java
@@ -22,17 +22,11 @@ import java.util.UUID;
 @Service
 @Slf4j
 public class PointService {
-    private final PaymentRepository paymentRepository;
 
     private final MemberRepository memberRepository;
 
-    private final PaymentEventHandler paymentEventHandler;
-
-
-    public PointService(PaymentRepository paymentRepository, MemberRepository memberRepository, PaymentEventHandler paymentEventHandler) {
-        this.paymentRepository = paymentRepository;
+    public PointService(MemberRepository memberRepository) {
         this.memberRepository = memberRepository;
-        this.paymentEventHandler = paymentEventHandler;
     }
 
     /**
@@ -72,76 +66,7 @@ public class PointService {
         return pointDto;
     }
 
-    @Transactional
-    public PointDto paid(UUID paymentId, Long amount) {
-        PointDto pointDto = new PointDto();
 
-
-        log.info(String.format(" ==== paid() start : request amount %d ", amount));
-
-        if (amount < 0) {
-            throw new IllegalArgumentException("Invalid amount");
-        }
-        Payment payment = paymentRepository.findById(paymentId).orElseThrow(() -> new InvalidAmountException(ErrorCode.INVALID_AMOUNT));
-
-        if (payment.getActuAmount() + amount == payment.getPrice()) { // 완납 경우
-            // 1. 결제 완료 처리
-            payment.setPayYn("Y");
-            payment.setActuAmount(payment.getActuAmount() + amount);
-
-            // 2. 포인트 차감 처리
-            Users member = memberRepository.findMember(payment.getReservation().getMember().getUserId()).orElseThrow(() -> new UserNotFoundException(ErrorCode.MEMBER_NOT_FOUND));
-            if (member.getUserPoint() <= 0) {
-                throw new InvalidAmountException(ErrorCode.NOT_ENOUGH_AMOUNT);
-            }
-            member.setUserPoint(member.getUserPoint() - amount);
-            memberRepository.save(member);
-
-            // 3. 자리 확정 처리
-            payment.getReservation().setConfirmYn("Y");
-            payment.getReservation().getSeat().setStatus(SeatEnum.OCCUPIED.getStatus());
-
-            paymentRepository.saveAndFlush(payment);
-
-            // 결제 정보 publish 로직 추가
-            PaymentEvent paymentEvent = PaymentEvent.builder()
-                            .payYn("Y")
-                                    .confirmYn("Y")
-                    .actuAmount(payment.getActuAmount() + amount)
-                    .status(SeatEnum.OCCUPIED.getStatus()).build();
-
-            paymentEventHandler.publish(paymentEvent);
-
-            // 4. return 값 생성
-            pointDto.setPoint(member.getUserPoint());
-            pointDto.setMemberId(member.getUserId());
-            pointDto.setConcertId(payment.getReservation().getSeat().getConcertSchedule().getConcertId().getConcertId());
-            log.info(String.format(" ==== paid() end : All paid result : %d ====", pointDto.getPoint()));
-
-        } else { // 분납인 경우
-            // 1.포인트 차감
-            Users member = payment.getReservation().getMember();
-            if (member.getUserPoint() <= 0) {
-                throw new InvalidAmountException(ErrorCode.NOT_ENOUGH_AMOUNT);
-            }
-            member.setUserPoint(member.getUserPoint() - amount);
-            memberRepository.save(member);
-
-            // 2.결제 완료 처리
-            payment.setActuAmount(payment.getActuAmount() + amount);
-            paymentRepository.saveAndFlush(payment);
-
-            // 결제 정보 publish 로직 추가
-            PaymentEvent paymentEvent = PaymentEvent.builder()
-                            .actuAmount(payment.getActuAmount() + amount).build();
-            paymentEventHandler.publish(paymentEvent);
-
-            // 4. return 값 생성
-            pointDto.setPoint(member.getUserPoint());
-            log.info(String.format(" ==== paid() end : Not all paid result : %d ====", pointDto.getPoint()));
-        }
-        return pointDto;
-    }
 
     /**
      * 사용자의 포인트 조회
