+++ src/main/java/com/hhplus/concertReserv/interfaces/controller/PointController.java
@@ -56,18 +56,5 @@ public class PointController {
         return ResponseEntity.ok().body(pointFacade.charge(requestBody));
     }
 
-    /**
-     * 임시배정된 자리를 결제
-     * 토큰 안에 만료시간, 콘서트 UUID, 사용자 UUID 가 있다.
-     * requestBody 안에 결제 UUID가 있다. 결제 식별자로 좌석 정보를 알 수 있다.
-     * @param requestBody requestBody
-     * @return 결제하고 나고 남은 포인트
-     */
-    @PostMapping("/paid")
-    @Operation(summary = "임시배정된 자리를 결제 요청", description = "임시 배정된 자리의 결제 처리 완료")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "Success", content = {@Content(schema = @Schema(implementation = PointDto.class))})
-            , @ApiResponse(responseCode = "404", description = "Not Found")})
-    public ResponseEntity<Object> paid(@RequestBody PointCommand.Paid requestBody) {
-        return ResponseEntity.ok().body(pointFacade.paid(requestBody));
-    }
+
 }
