+++ src/test/java/com/hhplus/concertReserv/domain/service/PointServiceTest.java
@@ -23,6 +23,7 @@ class PointServiceTest {
     private PointService pointService;
     @Mock
     private MemberRepository memberRepository;
+
     @DisplayName("충전금액이 음수인 경우 exception 처리")
     @Test
     void charge() {
@@ -34,16 +35,7 @@ class PointServiceTest {
         assertThat(pointService.charge(memberId, amount).getResult()).isEqualTo(false);
     }
 
-    @DisplayName("결제금액이 음수인 경우 exception 처리")
-    @Test
-    void paid() {
 
-        Long amount = -1L;
-        UUID memberId = UUID.fromString("550e8400-e29b-41d4-a716-446655440000");
-
-        //then
-        assertThat(pointService.paid(memberId, amount).getResult()).isEqualTo(false);
-    }
 
     /**
      * 등록되지 않은 사용자일 경우 exception 처리
